+++ src/app/account/additional-components/account-flows-tab/account-flows-subtabs/account-flows-datasets-subtab/components/custom-filters-view/custom-filters-view.component.spec.ts
@@ -34,6 +34,7 @@ describe("CustomFiltersViewComponent", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
+            applyFilters: false,
         };
     });
 
@@ -81,6 +82,7 @@ describe("CustomFiltersViewComponent", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
+            applyFilters: false,
         };
         fixture.detectChanges();
         component.onChangeNextAttemptFilter();
@@ -102,6 +104,7 @@ describe("CustomFiltersViewComponent", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
+            applyFilters: false,
         };
         fixture.detectChanges();
         emitClickOnElementByDataTestId(fixture, "clear-from-control");
@@ -123,6 +126,7 @@ describe("CustomFiltersViewComponent", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
+            applyFilters: false,
         };
         fixture.detectChanges();
         emitClickOnElementByDataTestId(fixture, "clear-to-control");
@@ -144,6 +148,7 @@ describe("CustomFiltersViewComponent", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
+            applyFilters: false,
         };
         fixture.detectChanges();
         emitClickOnElementByDataTestId(fixture, "clear-after-control");
@@ -165,6 +170,7 @@ describe("CustomFiltersViewComponent", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
+            applyFilters: false,
         };
         fixture.detectChanges();
         emitClickOnElementByDataTestId(fixture, "clear-before-control");
@@ -186,6 +192,7 @@ describe("CustomFiltersViewComponent", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
+            applyFilters: false,
         };
         fixture.detectChanges();
         component.onQuickRangeLastAttempt({
@@ -212,6 +219,7 @@ describe("CustomFiltersViewComponent", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
+            applyFilters: false,
         };
         fixture.detectChanges();
         component.onQuickRangeNextAttempt({