+++ src/app/account/additional-components/account-flows-tab/account-flows-subtabs/account-flows-datasets-subtab/components/triage-filters-view/triage-filters-view.component.spec.ts
@@ -33,6 +33,7 @@ describe("TriageFiltersViewComponent", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
+            applyFilters: false,
         };
     });