+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Lazy loading of cards when scrolling for account filters.
+
 ## [0.61.1] - 2025-12-10
 ### Fixed
 - Account flows tab: pagination bug for page more than 1